feat: implement database schema improvements and user management
- Add Foreign Key constraint to carts.user_id with ON DELETE CASCADE
- Convert Float to Numeric(10,2) for all monetary values (products.price, orders.total, order_items.price, order_items.subtotal)
- Add DELETE /users/{id} endpoint (admin only)
- Fix products search, filter, and sort functionality in repository
- Implement order status transition validation
- Add admin order listing endpoint
- Update health check to require admin auth for /health/db
- Prevent admin from deleting their own account

+++ app/modules/products/models.py
@@ -2,7 +2,7 @@ import uuid
 from datetime import datetime
 from typing import TYPE_CHECKING
 
-from sqlalchemy import Boolean, DateTime, Float, ForeignKey, Integer, String, Text, func
+from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, Numeric, String, Text, func
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -23,7 +23,7 @@ class Product(Base):
 
     name: Mapped[str] = mapped_column(String(255), nullable=False)
     description: Mapped[str] = mapped_column(Text, nullable=False)
-    price: Mapped[float] = mapped_column(Float, nullable=False)
+    price: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
     stock: Mapped[int] = mapped_column(Integer, nullable=False)
     category_id: Mapped[uuid.UUID] = mapped_column(
         UUID(as_uuid=True), ForeignKey("categories.id"), nullable=False, index=True
